requirements.txt
asgiref==3.8.1
attrs==25.3.0
black==25.1.0
click==8.1.8
colorama==0.4.6
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.7
psycopg==3.2.6
psycopg-binary==3.2.6
PyJWT==2.10.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
